Move EntityTile to components dir
