+++ context/app/static/js/components/EntityTile/style.js
